DP-107 Moved misplaced unit test classes
